requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
Django==2.0.3
django-configurations==2.0
django-dotenv==1.4.2
django-extensions==2.0.2
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
kombu==4.1.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
